power: supply: bq2597x_charger_pipa: fix void to int pointer cast

+++ drivers/power/supply/ti/bq2597x_charger_pipa.c
@@ -1500,9 +1500,9 @@ static int bq2597x_get_dev_role(struct i2c_client *client)
 	}
 
 	dev_info(&client->dev, "%s: matched to %s, dev_role: %d.\n",
-			__func__, of_id->compatible, (int)of_id->data);
+			__func__, of_id->compatible, (uintptr_t)of_id->data);
 
-	return (int)of_id->data;
+	return (uintptr_t)of_id->data;
 }
 
 static int bq2597x_parse_dt(struct bq2597x *bq, struct device *dev)
